requirements-for-testing.txt
azureml-telemetry
azureml-core
raiwidgets~=0.29.0
responsibleai~=0.29.0
nbformat
nbval
nteract-scrapbook
papermill
pytest
pytest-xdist
numpy<1.24.0
mlflow
azureml-rai-utils==0.0.5